requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
install==1.3.4
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytz==2021.1
SQLAlchemy==1.4.15
Werkzeug==2.0.1
